jobs/requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
paho-mqtt==1.6.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
requests==2.28.0
sendgrid==6.10.0
six==1.16.0
SQLAlchemy==1.4.37
starkbank-ecdsa==2.2.0
urllib3==1.26.9
